feat(ww3d2-gl): Milestone 4 Step 6 - Tests/RenderTexturePipeline, the exit criterion
Adds the real texture pipeline end-to-end harness (TextureClass/
TextureLoader/TextureFilterClass/MissingTexture/DDSFileClass/
BitmapHandlerClass), exercising 6 real pixel-level checks: foreground
TGA load, engine-generated mip chain, DDS/DXT1 software decode,
missing-texture fallback, a real cross-thread (pthread) background
texture load, and WRAP/CLAMP sampler addressing. This is Milestone 4's
exit criterion (docs/native-port-plan.md, Draft 22 Step 6).

Three genuine, real, pre-existing engine bugs were found and fixed
along the way - all only ever surfaced because this is the first time
real, non-uniform texture content has been loaded through this code
path on this platform:

- TARGA.h: TGA2Footer/TGA2Extension used `long` for on-disk 4-byte TGA
  2.0 fields (Extension, Developer, KeyColor, ColorCor, PostStamp,
  ScanLine). `long` is 8 bytes under LP64 (Linux/macOS 64-bit) but 4
  under Windows LLP64, so sizeof(TGA2Footer) silently grew from 26 to
  34 bytes, breaking the hardcoded File_Seek(-26, SEEK_END) footer
  read used by every real TGA load on this platform. Fixed by using
  int32_t, matching the format spec exactly.

- dx8wrapper_gl.cpp Create_Device(): the real Windows path populates
  D3DFormatToWW3DFormatConversionArray via Init_D3D_To_WW3_Conversion(),
  called once from WW3D::Init() (ww3d.cpp:189) before DX8Wrapper::Init()
  ever runs. Milestone 1's Trap 1 deliberately keeps GL's Create_Device()
  out of that monolithic init chain, so the call never happened - the
  array stayed zero-initialized (all WW3D_FORMAT_UNKNOWN), which made
  D3DFormat_To_WW3DFormat(DisplayFormat) return UNKNOWN, which made
  every DX8Caps::Support_Texture_Format() check fail (the
  display_format==WW3D_FORMAT_UNKNOWN early-exit in
  Check_Texture_Format_Support), silently collapsing every real texture
  load's format down Get_Valid_Texture_Format's fallback chain to a
  16-bit format regardless of the requested 32-bit bit depth. Fixed by
  calling the same portable formconv.cpp function from the GL-analogous
  spot in Create_Device().

- bitmaphandler.cpp Copy_Image(): the same-format (A8R8G8B8/X8R8G8B8)
  fast path's "copy current level while box-filtering the next mip
  level in place" branch only special-cased dest_surface_width==1 (the
  terminal 1x1 mip level of a square power-of-two chain). A non-square
  texture whose height has already bottomed out at 1 while width has
  not (e.g. a 2x1 texture) fell through to the general 2D box-filter
  loop, whose `dest_surface_height/2` bound integer-divides 1/2 to 0 -
  the loop never runs, and the destination surface is left as whatever
  garbage was already in its shadow buffer. Fixed by adding a symmetric
  dest_surface_height==1 case doing a horizontal-only pairing (no
  next-row read, since there is no next row).

Verification:
- WSL2 linux-x64 `-k 0` full-error-count build of g_gameenginedevice/
  z_gameenginedevice: exactly 17 errors each (34 total), per-file
  breakdown unchanged (atlbase 6, winsock 4, imagehlp 3, d3dx8math/
  BezierSegment 3, mbstring 1 - no WW3D2 file among them) - matches
  the established baseline exactly, zero regression.
- Real MSVC win32 build of g_ww3d2/z_ww3d2/g_gameenginedevice/
  z_gameenginedevice: 0 errors (only pre-existing, unrelated C4018
  warnings in ConnectionManager.cpp/W3DDisplay.cpp).
- All four GL harnesses rebuilt and re-run under WSLg DISPLAY=:0:
  RenderDeviceInitTest, RenderTexturedTriangleTest,
  RenderEngineDrawPathTest bit-identical to prior runs;
  RenderTexturePipelineTest passes all 6 checks, stable across 5
  consecutive runs (the pre-fix run was nondeterministic - background-
  thread timing changed which uninitialized bytes got read).
- Wired into .github/workflows/linux-native.yml, mirroring the
  existing build+run+summary pattern for the other three harnesses.

+++ Core/Libraries/Source/WWVegas/WWLib/TARGA.h
@@ -18,6 +18,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 /****************************************************************************
 *
 *         C O N F I D E N T I A L --- W E S T W O O D   S T U D I O S
@@ -133,8 +135,16 @@ typedef struct _TGAHeader
  */
 typedef struct _TGA2Footer
 	{
-	long Extension;
-	long Developer;
+	// native port plan Phase 2 (64-bit pointer-width cleanliness), found by
+	// Milestone 4 Step 6's harness actually loading a real TGA for the first
+	// time on 64-bit Linux: `long` is 8 bytes under LP64 (Linux/macOS 64-bit)
+	// but 4 bytes under Windows' LLP64 - these two fields are genuinely
+	// on-disk TGA 2.0 format data (4-byte file offsets), and sizeof(TGA2Footer)
+	// being fed into a hardcoded-26-byte seek/read pair silently grew to 34
+	// bytes, making every real TGA load fail with TGAERR_READ on this
+	// platform. int32_t is exactly what the format specifies everywhere.
+	int32_t Extension;
+	int32_t Developer;
 	char Signature[16];
 	char RsvdChar;
 	char BZST;
@@ -224,12 +234,12 @@ typedef struct _TGA2Extension
 	TGA2TimeStamp JobTime;
 	char          SoftID[41];
 	TGA2SoftVer   SoftVer;
-	long          KeyColor;
+	int32_t       KeyColor; // on-disk TGA 2.0 format data - see TGA2Footer's comment
 	TGA2Ratio     Aspect;
 	TGA2Ratio     Gamma;
-	long          ColorCor;
-	long          PostStamp;
-	long          ScanLine;
+	int32_t       ColorCor;
+	int32_t       PostStamp;
+	int32_t       ScanLine;
 	char          Attributes;
 	} TGA2Extension;
